local-pv-provisioner: support k8s v1.27
This commit adds support for k8s v1.27 to local-pv-provisioner. For
that, controller-runtime has been upgraded to v0.15.2.

From controller-runtime v0.15.0 the first argument of Builder.Watch()
method has changed from source.Source to client.Object, which leads a
compile error of local-pv-provisioner. To address this issue, we use
WatchRawSources() instead which behaves as the former Watch() method.